model_code/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
dnslib==0.9.23
dnspython==2.0.0
filelock==3.12.2
idna==3.4
ipwhois==1.2.0
joblib==1.3.1
numpy==1.25.0
pandas==2.0.3
py-radix==0.10.0
pyasn==1.6.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-file==1.5.1
scikit-learn==1.0.2
scipy==1.11.1
six==1.16.0
threadpoolctl==3.1.0
tldextract==3.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
whoisit==2.7.2
